backend/requirements.txt
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
numpy==1.26.0
deap==1.4.1